Extract crud-resource mixin
